// src/app/interfaces/placedto.ts
export interface PlaceDTO {

    id?: number;

    name: string

    city: string

    country: string

    placeCategory: string

    age: number

    description: string

    countLike?: number;

    countDisLike?: number;

    visitingHours: string

    thicketPrice: string

    bestTimeForVisit: string

    webSite: string

    status?: string;

    imageName?: string;

}